refactor: strictly enforce np.float32 casting for MS arrays
- Updates `src/MassFlow/processing.py` to cast array items to float32 at the initial ingestion phase in `process_spectra_batch`.
- Updates `src/MassFlow/database.py` to ensure schema legacy arrays and storage serialization strictly default to float32.
- Implements `Float32` custom type validation inside `src/MassFlow/models.py` Pydantic models for array annotations.
- Adjusts strict precision checking tests using `pytest.approx()` and extended tolerances to pass safely under standard `np.float32` boundaries.

+++ tests/test_mathematical_proof.py
@@ -25,8 +25,8 @@ def _make_spectrum(
 ) -> Spectrum:
     """Build a matchms.Spectrum with the mandatory metadata fields."""
     return Spectrum(
-        mz=np.array(mz, dtype=np.float64),
-        intensities=np.array(intensities, dtype=np.float64),
+        mz=np.array(mz, dtype=np.float32),
+        intensities=np.array(intensities, dtype=np.float32),
         metadata={
             "precursor_mz": precursor_mz,
             "id": spec_id,
